refactor : status를 card에서 분리
- card에서 분리했습니다.
