+++ backend/src/main/java/com/codeSquad/cocokyu/domain/model/Status.java
@@ -1,0 +1,9 @@
+package com.codeSquad.cocokyu.domain.model;
+
+
+public enum Status {
+    TODO,
+    DOING,
+    DONE,
+    DELETED
+}
